[Change] consolidate general stabilizer tests
Skipping tests for RotatedPlanar3DCode which are failing for now.
Will fix later.
